support_chatbot_for_your_website/requirements.txt
langchain
langchain-openai
langchain-community
langchain-pinecone
langchain-text-splitters
streamlit
lxml
beautifulsoup4
python-dotenv